UserAccessChecker-Python-main/requirements.txt
azure-functions
azure-cosmos>=4.5.0
azure-identity>=1.14.0
PyJWT>=2.8.0
cryptography>=41.0.0
requests>=2.31.0
